Make credit balances and ledger entries precise to two decimals
The gateway previously rounded all credits to integers, which made small
inference charges coarse and reduced ledger fidelity. Credits, manual
adjustments, and inference deductions now preserve two decimal places across the
backend and admin UI.

Constraint: Keep SQLite as the default local store while improving credit
precision without a full database replacement.
Rejected: Keep integer credits and only pretty-print decimals in the UI | that
would leave settlement behavior imprecise and misleading.
Confidence: high
Scope-risk: moderate
Directive: Treat credits as two-decimal values everywhere user-visible or
ledger-visible; do not reintroduce integer-only billing assumptions.
Tested: ./backend/.venv/bin/python -m pytest -q backend/tests/test_credits.py backend/tests/test_internal_public_gateway.py; cd frontend && npm test -- --run src/App.test.tsx && npm run build
Not-tested: Migration of very old existing local databases with hand-edited credit values

+++ backend/app/api/admin.py
@@ -148,12 +148,12 @@ class AccountRead(BaseModel):
     id: int
     name: str
     status: str
-    credit_balance: int
+    credit_balance: float
     notes: str | None = None
 
 
 class CreditAdjustmentCreate(BaseModel):
-    credits_delta: int
+    credits_delta: float
     notes: str | None = None
 
 
@@ -163,8 +163,8 @@ class CreditLedgerRead(BaseModel):
     api_key_id: int | None = None
     usage_record_id: int | None = None
     entry_type: str
-    credits_delta: int
-    balance_after: int
+    credits_delta: float
+    balance_after: float
     usd_amount: float | None = None
     provider_name: str | None = None
     model_id: str | None = None
